+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - (Fix) Metrics: throw exception during AUC calculation if metric has negative values
 - (New) Metrics: add parameter to optionally calculate natural log of metric AUC.
 - (New) Logging: improve default logging configuration
+- (New) Add Algorithm::setName to allow changing the algorithm name after creation, util for example for algorithms built by the autoconfig components
 
 # v0.18
 - (New) Property `solver.max-derivation-repetition` allows restricting algorithms generated from the grammar by